requirements.txt
azureml-core
azure-cli
